tokio_tungstenite

Module connect

Connection helper.

Functions

connect 🔒

connect_async

Connect to a given URL.

connect_async_tls_with_config

The same as connect_async() but the one can specify a websocket configuration, and a TLS connector to use. Please refer to connect_async() for more details. disable_nagle specifies if the Nagle’s algorithm must be disabled, i.e. set_nodelay(true). If you don’t know what the Nagle’s algorithm is, better leave it to false.

connect_async_with_config

The same as connect_async() but the one can specify a websocket configuration. Please refer to connect_async() for more details. disable_nagle specifies if the Nagle’s algorithm must be disabled, i.e. set_nodelay(true). If you don’t know what the Nagle’s algorithm is, better leave it set to false.