Module tokio_tungstenite::tls

Connection helper.

Modules

encryption

Encrypt a stream usin Tls

Enums

Connector

A connector that can be used when establishing connections, allowing to control whether native-tls or rustls is used to create a TLS connection. Or TLS can be disabled with the Plain variant.

MaybeTlsStream

A stream that might be protected with TLS.

Functions

client_async_tls

Creates a WebSocket handshake from a request and a stream, upgrading the stream to TLS if required.

client_async_tls_with_config

The same as client_async_tls() but the one can specify a websocket configuration, and an optional connector. If no connector is specified, a default one will be created.

tcp_tls

Given a domain and mode